Move setup_metrics, setup_tracing to coyote-server
Doesn't need to be part of the coyote lib crate.
