+++ server/Cargo.toml
@@ -10,8 +10,14 @@ clap.workspace = true
 coyote.workspace = true
 dotenvy.workspace = true
 mimalloc.workspace = true
+opentelemetry.workspace = true
+opentelemetry-otlp.workspace = true
+opentelemetry-semantic-conventions.workspace = true
+opentelemetry_sdk.workspace = true
 reqwest.workspace = true
 tokio.workspace = true
+tracing.workspace = true
+tracing-opentelemetry.workspace = true
 tracing-subscriber.workspace = true
 
 [lints]
